Refactor/remove duplicated-code
