+++ src/PushNotification/Message/Strategy/AndroidMessages.php
@@ -3,8 +3,7 @@
 
 namespace PushNotification\Message\Strategy;
 
-use PushNotification\Message\AppleMessage;
-use PushNotification\Message\BasicMessageInterface;
+use PushNotification\Message\BasicMessageAbstract;
 use PushNotification\Message\GoogleMessage;
 use PushNotification\Message\Config\Provider;
 
@@ -12,100 +11,12 @@ use PushNotification\Message\Config\Provider;
  * Class AndroidMessages
  * @package PushNotification\Message\Strategy
  */
-class AndroidMessages implements BasicMessageInterface, GoogleMessage
+class AndroidMessages extends BasicMessageAbstract implements GoogleMessage
 {
-
-    use Provider;
-
-    /** @var  string */
-    private $title;
-
-    /** @var  string */
-    private $body;
-
-    /** @var */
-    private $data;
-
-    /** @var  string related action like : events, coins,... */
-    private $action;
-
-    /** @var  string */
-    private $provider;
-
-    /** @var  array of users which targeted */
-    private $targets;
 
     public function __construct()
     {
         $this->setProvider(Provider::google());
-    }
-
-    /**
-     * @return string
-     */
-    public function getProvider()
-    {
-        return $this->provider;
-    }
-
-    /**
-     * @param string $provider
-     * @return AndroidMessages
-     */
-    public function setProvider($provider)
-    {
-        $this->provider = $provider;
-        return $this;
-    }
-
-    /**
-     * @param mixed $targets
-     * @return AndroidMessages
-     */
-    public function setTargets($targets)
-    {
-        $this->targets = $targets;
-        return $this;
-    }
-
-    /**
-     * @param string $action
-     * @return AndroidMessages
-     */
-    public function setAction($action)
-    {
-        $this->action = $action;
-        return $this;
-    }
-
-    /**
-     * @param mixed $data
-     * @return AndroidMessages
-     */
-    public function setData($data)
-    {
-        $this->data = $data;
-        return $this;
-    }
-
-    /**
-     * @param string $title
-     * @return AndroidMessages
-     */
-    public function setTitle($title)
-    {
-        $this->title = $title;
-        return $this;
-    }
-
-    /**
-     * @param string $body
-     * @return AndroidMessages
-     */
-    public function setBody($body)
-    {
-        $this->body = $body;
-        return $this;
     }
 
     /**
